Rework getCurrentUrl to not need request in template
