+++ modules/viewGlobals.js
@@ -1,53 +1,9 @@
 'use strict';
 
 const config = require('config');
-const { URL } = require('url');
-const querystring = require('querystring');
 const shortid = require('shortid');
 
-const { getBaseUrl, isWelsh, makeWelsh, removeWelsh, stripTrailingSlashes } = require('./urls');
 const formHelpers = require('./forms');
-
-/**
- * getCurrentUrl
- * Look up the current URL and rewrite to another locale
- */
-function getCurrentUrl(req, requestedLocale) {
-    const urlPath = req.originalUrl;
-    const baseUrl = getBaseUrl(req);
-
-    const isCurrentUrlWelsh = isWelsh(urlPath);
-    const isCyWithEnRequested = isCurrentUrlWelsh && requestedLocale === 'en';
-    const isEnWithCyRequested = !isCurrentUrlWelsh && requestedLocale === 'cy';
-
-    // Rewrite URL to requested language
-    let urlPathForRequestedLocale = urlPath;
-    if (isEnWithCyRequested) {
-        urlPathForRequestedLocale = makeWelsh(urlPath);
-    } else if (isCyWithEnRequested) {
-        urlPathForRequestedLocale = removeWelsh(urlPath);
-    }
-
-    // Remove any trailing slashes (eg. /welsh/ => /welsh)
-    const cleanedUrlPath = stripTrailingSlashes(urlPathForRequestedLocale);
-
-    const fullUrl = baseUrl + cleanedUrlPath;
-
-    const parsedUrl = new URL(fullUrl);
-    const parsedPathname = parsedUrl.pathname;
-    const parsedQuery = parsedUrl.search.replace(/^\?/, '');
-
-    // Remove draft and version parameters
-    const originalQuery = querystring.parse(parsedQuery);
-    delete originalQuery.version;
-    delete originalQuery.draft;
-
-    // Reconstruct clean URL
-    const newCleanQuery = querystring.stringify(originalQuery);
-    const newCleanUrl = newCleanQuery.length > 0 ? `${parsedPathname}?${newCleanQuery}` : parsedPathname;
-
-    return newCleanUrl;
-}
 
 function getCurrentSection(sectionId, pageId) {
     const isHomepage = sectionId === 'toplevel' && pageId === 'home';
@@ -66,8 +22,6 @@ function init(app) {
     setViewGlobal('shortid', () => shortid());
 
     setViewGlobal('anchors', config.get('anchors'));
-
-    setViewGlobal('getCurrentUrl', getCurrentUrl);
 
     setViewGlobal('getCurrentSection', getCurrentSection);
 
@@ -96,6 +50,5 @@ function init(app) {
 
 module.exports = {
     init,
-    getCurrentUrl,
     getCurrentSection
 };
