+++ views/components/preview.njk
@@ -3,7 +3,7 @@
         <div class="preview-banner">
             ✋ You are viewing a draft (last updated {{ status.lastUpdated }}).
             Please do not share this page.
-            <a href="{{ getCurrentUrl(request) }}">View original</a>
+            <a href="{{ getCurrentUrl() }}">View original</a>
         </div>
     {% endif %}
 {% endmacro %}
