+++ views/layouts/profiles.njk
@@ -31,7 +31,7 @@
             </div>
             <div class="page-section__supplementary accent--{{ pageAccent }} a--border-top">
                 <aside class="content-box">
-                    {{ nestedMenu(navigation, getCurrentUrl(request)) }}
+                    {{ nestedMenu(navigation, getCurrentUrl()) }}
                 </aside>
 
                 {% set quickLinksList = [] %}
